trend_engine/requirements.txt
requests
beautifulsoup4
google-genai
python-dotenv
supabase
playwright
lxml
